Added a stream capture locking feature due to DLive spuriously reporting a streamer as not live when they are and causing duplicate captures.
It works by having the capture script touch a file before it begins capturing, then remove the file when the capture is complete.

The bot will check if this file is present before checking if a DLive streamer is actually live which will reduce the amount of API hits and prevent it from going live twice.